+++ lib/fittor_store/index.dart
@@ -2,6 +2,9 @@ library;
 
 export 'fittor_store.dart';
 export 'fittor_store_mixin.dart';
-export 'persistent_storage.dart';
+export 'mobile_storage.dart';
 export 'secure_storage.dart';
 export 'security_manager.dart';
+export 'storage_factory.dart';
+export 'storage_interface.dart';
+export 'web_storage.dart';